+++ Delta.CertXplorer/src/Delta.CapiNet/Asn1/Asn1ObjectFactory.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Linq;
 using Delta.CapiNet.Logging;
 
 namespace Delta.CapiNet.Asn1
@@ -8,11 +9,10 @@ namespace Delta.CapiNet.Asn1
         private static ILogService log = LogManager.GetLogger(typeof(Asn1ObjectFactory));
 
         public Asn1Object CreateAsn1Object(
-            Asn1Document document, 
-            TaggedObject content, 
+            Asn1Document document,
+            TaggedObject content,
             Asn1Object parent)
         {
-            Asn1Object result = null;
             try
             {
                 if (document.ShowInvalidTaggedObjects && IsInvalidTaggedObject(content))
@@ -26,45 +26,53 @@ namespace Delta.CapiNet.Asn1
                 }
 
                 var asn1TagValue = tagValue.GetAsn1TagValue();
-                switch (asn1TagValue)
-                {
-                    case (int)Asn1Tags.Boolean:
-                        return new Asn1Boolean(document, content, parent);
-                    case (int)Asn1Tags.Integer:
-                        return new Asn1Integer(document, content, parent);
-                    case (int)Asn1Tags.BitString:
-                        return new Asn1BitString(document, content, parent);
-                    case (int)Asn1Tags.OctetString:
-                        return new Asn1OctetString(document, content, parent);
-                    case (int)Asn1Tags.Null:
-                        return new Asn1Null(document, content, parent);
-                    case (int)Asn1Tags.ObjectIdentifier:
-                        return new Asn1Oid(document, content, parent);
-                    case (int)Asn1Tags.Utf8String:
-                        return new Asn1Utf8String(document, content, parent);
-                    case (int)Asn1Tags.Sequence:
-                        return new Asn1Sequence(document, content, parent);
-                    case (int)Asn1Tags.Set:
-                        return new Asn1Set(document, content, parent);
-                    case (int)Asn1Tags.NumericString:
-                        return new Asn1NumericString(document, content, parent);
-                    case (int)Asn1Tags.PrintableString:
-                        return new Asn1PrintableString(document, content, parent);
-                    case (int)Asn1Tags.UtcTime:
-                        return new Asn1UtcTime(document, content, parent);
-                    default:
-                        var specific = CreateSpecificTaggedObject(document, content, parent);                        
-                        return specific ?? CreateUnsupportedTaggedObject(document, content, parent);
-                }
+                if (IsAsn1Tag(asn1TagValue))
+                    return CreateUniversalTaggedObject((Asn1Tag)asn1TagValue, document, content, parent);
+
+                // Otherwise
+                return CreateSpecificOrUnsupportedTaggedObject(document, content, parent);
             }
             catch (Exception ex)
             {
                 // Could not create a "real" ASN1 object --> return the special "Unsupported" object.
                 log.Error(ex);
-                result = CreateUnsupportedTaggedObject(document, content, parent, ex);
+                return CreateUnsupportedTaggedObject(document, content, parent, ex);
+            }
+        }
+
+        protected Asn1Object CreateUniversalTaggedObject(Asn1Tag asn1Tag, Asn1Document document, TaggedObject content, Asn1Object parent)
+        {
+            switch (asn1Tag)
+            {
+                case Asn1Tag.Boolean:
+                    return new Asn1Boolean(document, content, parent);
+                case Asn1Tag.Integer:
+                    return new Asn1Integer(document, content, parent);
+                case Asn1Tag.BitString:
+                    return new Asn1BitString(document, content, parent);
+                case Asn1Tag.CharacterString:
+                    return new Asn1CharacterString(document, content, parent);
+                case Asn1Tag.OctetString:
+                    return new Asn1OctetString(document, content, parent);
+                case Asn1Tag.Null:
+                    return new Asn1Null(document, content, parent);
+                case Asn1Tag.ObjectIdentifier:
+                    return new Asn1Oid(document, content, parent);
+                case Asn1Tag.Utf8String:
+                    return new Asn1Utf8String(document, content, parent);
+                case Asn1Tag.Sequence:
+                    return new Asn1Sequence(document, content, parent);
+                case Asn1Tag.Set:
+                    return new Asn1Set(document, content, parent);
+                case Asn1Tag.NumericString:
+                    return new Asn1NumericString(document, content, parent);
+                case Asn1Tag.PrintableString:
+                    return new Asn1PrintableString(document, content, parent);
+                case Asn1Tag.UtcTime:
+                    return new Asn1UtcTime(document, content, parent);                
             }
 
-            return result;
+            return null;
         }
 
         protected virtual Asn1Object CreateSpecificTaggedObject(Asn1Document document, TaggedObject content, Asn1Object parent)
@@ -72,7 +80,12 @@ namespace Delta.CapiNet.Asn1
             return null;
         }
 
-        private Asn1Object CreateUnsupportedTaggedObject(Asn1Document document, TaggedObject content, Asn1Object parent, Exception exception = null)
+        protected bool IsAsn1Tag(int tagValue)
+        {
+            return Enum.GetValues(typeof(Asn1Tag)).Cast<Asn1Tag>().Select(t => (int)t).Any(v => v == tagValue);
+        }
+
+        protected Asn1Object CreateUnsupportedTaggedObject(Asn1Document document, TaggedObject content, Asn1Object parent, Exception exception = null)
         {
             if (exception != null)
                 return new Asn1Unsupported(document, content, parent, exception);
@@ -82,6 +95,12 @@ namespace Delta.CapiNet.Asn1
                 return new Asn1ContextSpecific(document, content, parent);
 
             return new Asn1Unsupported(document, content, parent);
+        }
+
+        private Asn1Object CreateSpecificOrUnsupportedTaggedObject(Asn1Document document, TaggedObject content, Asn1Object parent)
+        {
+            var specific = CreateSpecificTaggedObject(document, content, parent);
+            return specific ?? CreateUnsupportedTaggedObject(document, content, parent);
         }
 
         private static bool IsInvalidTaggedObject(TaggedObject content)