+++ Delta.CertXplorer/src/Delta.CapiNet/Properties/AssemblyInfo.cs
@@ -15,5 +15,5 @@ using System.Runtime.InteropServices;
 
 internal static class AssemblyInfo
 {
-    public const string Version = "1.2.3.0";
+    public const string Version = "1.2.4.0";
 }