+++ Delta.CertXplorer/src/Delta.CertXplorer/Asn1Decoder/Asn1DocumentControl.cs
@@ -9,6 +9,8 @@ namespace Delta.CertXplorer.Asn1Decoder
 {
     internal partial class Asn1DocumentControl : UserControl
     {
+        private bool eventsSuspended = false;
+
         /// <summary>
         /// Initializes a new instance of the <see cref="Asn1DocumentControl"/> class.
         /// </summary>
@@ -18,15 +20,12 @@ namespace Delta.CertXplorer.Asn1Decoder
 
             ThemesManager.RegisterThemeAwareControl(this, (renderer) =>
             {
-                    if (renderer is ToolStripProfessionalRenderer)
-                        ((ToolStripProfessionalRenderer)renderer).RoundedEdges = false;
-                    tstrip.Renderer = renderer;
+                if (renderer is ToolStripProfessionalRenderer)
+                    ((ToolStripProfessionalRenderer)renderer).RoundedEdges = false;
+                tstrip.Renderer = renderer;
             });
 
             tstrip.SetRoundedEdges(false);
-            parseOctetStringsToolStripButton.Checked = viewer.ParseOctetStrings;
-            showInvalidTaggedObjectsToolStripButton.Checked = viewer.ShowInvalidTaggedObjects;
-            parseIcaoMrtdToolStripButton.Checked = viewer.IsIcaoMrtd;
         }
 
         public void SetData(byte[] bytes)
@@ -45,26 +44,61 @@ namespace Delta.CertXplorer.Asn1Decoder
         protected override void OnLoad(EventArgs e)
         {
             base.OnLoad(e);
-
-            //parseOctetStringsToolStripButton.Checked = false;
+            
             refreshToolStripButton.Click += (s, _) => viewer.ParseData();
+
+            parseOctetStringsToolStripButton.Checked = viewer.ParseOctetStrings;
             parseOctetStringsToolStripButton.CheckedChanged += (s, _) =>
             {
                 viewer.ParseOctetStrings = parseOctetStringsToolStripButton.Checked;
                 viewer.ParseData();
             };
 
-            parseIcaoMrtdToolStripButton.CheckedChanged += (s, _) =>
-            {
-                viewer.IsIcaoMrtd = parseIcaoMrtdToolStripButton.Checked;
-                viewer.ParseData();
-            };
-
+            showInvalidTaggedObjectsToolStripButton.Checked = viewer.ShowInvalidTaggedObjects;
             showInvalidTaggedObjectsToolStripButton.CheckedChanged += (s, _) =>
             {
                 viewer.ShowInvalidTaggedObjects = showInvalidTaggedObjectsToolStripButton.Checked;
                 viewer.ParseData();
             };
+
+            // Initialize parse mode
+
+            icaoToolStripMenuItem.Checked = false;
+            cvToolStripMenuItem.Checked = false;
+            standardToolStripMenuItem.Checked = false;
+
+            switch (viewer.ParseMode)
+            {
+                case Asn1ParseMode.Icao: icaoToolStripMenuItem.Checked = true; break;
+                case Asn1ParseMode.CardVerifiable: cvToolStripMenuItem.Checked = true; break;
+                default: standardToolStripMenuItem.Checked = true; break;
+            }
+
+            icaoToolStripMenuItem.CheckedChanged += (s, _) => UpdateParseMode(Asn1ParseMode.Icao);
+            cvToolStripMenuItem.CheckedChanged += (s, _) => UpdateParseMode(Asn1ParseMode.CardVerifiable);
+            standardToolStripMenuItem.CheckedChanged += (s, _) => UpdateParseMode(Asn1ParseMode.Standard);            
+        }
+
+        private void UpdateParseMode(Asn1ParseMode mode)
+        {
+            if (eventsSuspended) return;
+            eventsSuspended = true;
+
+            icaoToolStripMenuItem.Checked = false;
+            cvToolStripMenuItem.Checked = false;
+            standardToolStripMenuItem.Checked = false;
+
+            switch (mode)
+            {
+                case Asn1ParseMode.Icao: icaoToolStripMenuItem.Checked = true; break;
+                case Asn1ParseMode.CardVerifiable: cvToolStripMenuItem.Checked = true; break;
+                default: standardToolStripMenuItem.Checked = true; break;
+            }
+
+            eventsSuspended = false;
+
+            viewer.ParseMode = mode;
+            viewer.ParseData();
         }
     }
 }