+++ Delta.CertXplorer/History.md
@@ -1,7 +1,13 @@
 CertXplorer Versions History
 ============================
 
+Version 2.4.14 (2015/08/10)
+--------------------------
+* Delta.CapiNet: Version 1.2.4
+* Added first support for Card Verifiable documents (such as Certificates used in EAC)
+
 Version 2.4.13 (2014/11/18)
+--------------------------
 * Delta.CapiNet: Version 1.2.3
 * First code aimed at decoding ICAO/MRTD data